refactor: Refactor all city list related functions out into a new file
The functions related to the rendering of the autocomplete city list
have been moved into a new file cityViewUpdater.
